initial commit
All work up until now has been squashed into this commit.

This program reads from a MQTT protocol for updates. When an update
is received it inserts rows into a separate database. The credentials
for these two forms of communication are pulled from environment
variables.

This project was built with maven.